Fix most backend alignment issues
The remaining issue is the lack of `full_name` in the sign up route.
This is due to the OpenAPI spec not knowing about what data the
downstream `on_new_identity` callback uses to enrich it's user model.
This is a minor annoyance for now, but will be some friction.